# PeculiarPlaces/doc/openapi.yml
openapi: 3.0.0
info:
  title: PeculiarPlaces API
  description: API for discovering, sharing, and reviewing peculiar or interesting places.
  version: 1.0.0
servers:
  - url: /api
    description: Local development server
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from POST /api/users/
  schemas:
    PlaceInput:
      type: object
      properties:
        name:
          type: string
          example: "Mysterious Cave"
        description:
          type: string
          example: "A hidden cave with unusual rock formations"
        category:
          type: string
          example: "nature"
        address:
          type: string
          example: "Forest Road 1"
        postal_code:
          type: string
          example: "00100"
        city:
          type: string
          example: "Helsinki"
        latitude:
          type: number
          format: float
          example: 60.1699
        longitude:
          type: number
          format: float
          example: 24.9384
        application:
          type: string
          example: "MyApp"
      required:
        - name
        - latitude
        - longitude
...
    Place:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        name:
          type: string
          example: "Mysterious Cave"
        description:
          type: string
          example: "A hidden cave with unusual rock formations"
        category:
          type: string
          example: "nature"
        address:
          type: string
          example: "Forest Road 1"
        postal_code:
          type: string
          example: "00100"
        city:
          type: string
          example: "Helsinki"
        latitude:
          type: number
          format: float
          example: 60.1699
        longitude:
          type: number
          format: float
          example: 24.9384
        application:
          type: string
          example: "MyApp"
        trust_score:
          type: number
          format: float
          example: 4.0
...
paths:
...
  /places/{place_id}/:
    parameters:
      - name: place_id
        in: path
        required: true
        schema:
          type: integer
          example: 1

    get:
      summary: Get a single place
      responses:
        '200':
          description: Place details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
              example:
                id: 1
                user_id: 1
                name: "Mysterious Cave"
                description: "A hidden cave with unusual rock formations"
                category: "nature"
                address: "Forest Road 1"
                postal_code: "00100"
                city: "Helsinki"
                latitude: 60.1699
                longitude: 24.9384
                application: "MyApp"
                trust_score: 4.0
        '404':
          description: Place not found
          content:
            application/json:
              example:
                description: "Place not found"

    put:
      summary: Update a place (owner only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceInput'
            example:
              description: "Updated description"
              city: "Espoo"
      responses:
        '200':
          description: Place updated successfully
          content:
            application/json:
              example:
                message: "Place updated successfully"
        '400':
          description: Invalid field values
          content:
            application/json:
              example:
                error: "Invalid field values"
        '401':
          description: Authentication required
          content:
            application/json:
              example:
                error: "Authentication required"
        '403':
          description: Not the owner of this place
          content:
            application/json:
              example:
                error: "You do not own this resource"
        '404':
          description: Place not found
          content:
            application/json:
              example:
                description: "Place not found"
        '415':
          description: Content type must be application/json
          content:
            application/json:
              example:
                error: "Request content type must be JSON"

    delete:
      summary: Delete a place (owner only)
      description: Deletes the place and all associated reviews, images, and reports.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Place deleted successfully
          content:
            application/json:
              example:
                message: "Place deleted successfully"
        '401':
          description: Authentication required
          content:
            application/json:
              example:
                error: "Authentication required"
        '403':
          description: Not the owner of this place
          content:
            application/json:
              example:
                error: "You do not own this resource"
        '404':
          description: Place not found
          content:
            application/json:
              example:
                description: "Place not found"

  # ── Reviews ────────────────────────────────────────────────────────────────
...